+++ addon/locales/ru/translations.js
@@ -17,14 +17,7 @@ export default {
 
     'new-platform-flexberry-services-lock-list': {
       caption: 'Список блокировок',
-    },
-
-    'new-platform-flexberry-services-lock-edit': {
-      caption: 'Объект блокировки',
-      'lockKey-caption': 'Ключ заблокированного объекта',
-      'userName-caption': 'Заблокировавший пользователь',
-      'lockDate-caption': 'Дата блокировки',
-    },
+    }
   },
 
   'components': {